Ajuste final del proyecto2
Se hicieron los respectivos ajustes para que se logre observar de mejor manera la variacion de la temperatura y la variación de lux

+++ K32L2B31A_Project/iot_sdk/iot_sdk_peripherals/sensor_de_Temperatura.c
@@ -64,9 +64,9 @@ void SensorDeTemperaturaEsperarResultado(void){
 
 
 
-uint32_t SensorDeTemperaturaObtenerDatoADC(void){
+float SensorDeTemperaturaObtenerDatoADC(void){
 
-	uint32_t resultadoADC;
+	float resultadoADC;
 	SensorDeTemperaturaIniciarCaptura();
 	SensorDeTemperaturaEsperarResultado();
 	resultadoADC=ADC16_GetChannelConversionValue(SENSOR_DE_TEMPERATURA_ADC16_BASE, SENSOR_DE_TEMPERATURA_ADC16_CHANNEL_GROUP);
