+++ K32L2B31A_Project/iot_sdk/iot_sdk_peripherals/sensor_de_Temperatura.h
@@ -44,7 +44,7 @@
  * @brief Retorna resultado de la conversion ADC para sensor de Temperatura
  *
  */
-uint32_t SensorDeTemperaturaObtenerDatoADC(void);
+float SensorDeTemperaturaObtenerDatoADC(void);
 
 /** @} */ // end of SENSOR_DE_TEMPERATURA group
 /** @} */ // end of PERIPHERALS group
